move includes_image_modification to io/config
